requirements.txt
# CMake==3.6.3
numpy==1.17.4
pandas==0.25.3
jupyter==1.0.0
pytest==3.2.1
opencv-python==3.4.2.16
matplotlib==3.1.2
face_recognition==1.2.3
stopwatch.py==1.0.1
pyspark==2.4.3
findspark==1.3.0
torch==1.4.0
torchvision==0.5.0
facenet-pytorch==2.0.1
mtcnn==0.0.9
tensorflow-gpu==1.15.0
scikit-image==0.16.2
moviepy==1.0.1
librosa==0.7.2
fastai==1.0.59
# pyarrow==0.16.0